requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.3
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
-e git+https://github.com/django/django.git@62d85a283500e9abb0e1c9ec53c59be468f056a0#egg=Django
django-crispy-forms==1.9.1
gcloud==0.17.0
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.7.0
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpcio==1.27.2
httplib2==0.18.1
idna==2.10
isort==4.3.21
jws==0.1.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.0
oauth2client==3.0.0
Pillow==7.2.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycryptodome==3.4.3
pylint==2.5.2
pylint-django==2.0.15
pylint-plugin-utils==0.6
Pyrebase==3.0.27
python-jwt==2.0.1
pytz==2020.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
wrapt==1.12.1
